+++ config/webpack.config.js
@@ -3,6 +3,7 @@ const HtmlWebpackPlugin = require("html-webpack-plugin");
 const WebpackManifestPlugin = require("webpack-manifest-plugin");
 const { CleanWebpackPlugin } = require("clean-webpack-plugin");
 const { BundleAnalyzerPlugin } = require("webpack-bundle-analyzer");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 
 const output = path.resolve(__dirname, "..", "build");
 const input = path.resolve(__dirname, "..", "src", "index.js");
@@ -14,14 +15,65 @@ module.exports = (env, argv) => {
         output: {
             filename: "[name].[contenthash].js",
             path: output,
+            publicPath: "/",
         },
         module: {
             rules: [
                 {
-                    test: /\.(js|jsx)$/,
+                    test: /\.(js|jsx)$/i,
                     exclude: /node_modules/,
                     use: {
                         loader: "babel-loader",
+                    },
+                },
+                {
+                    test: /\.s[ac]ss$/i,
+                    use: [
+                        "style-loader",
+                        MiniCssExtractPlugin.loader,
+                        "css-loader",
+                        {
+                            loader: "sass-loader",
+                            options: {
+                                sassOptions: {
+                                    includePaths: [
+                                        path.resolve(__dirname, "..", "src"),
+                                    ],
+                                },
+                            },
+                        },
+                    ],
+                },
+                {
+                    test: /\.(woff2?|ttf|otf|eot|svg)$/,
+                    include: path.resolve(
+                        __dirname,
+                        "..",
+                        "src",
+                        "assets",
+                        "styles",
+                        "fonts"
+                    ),
+                    exclude: /node_modules/,
+                    use: {
+                        loader: "file-loader",
+                        options: {
+                            name: "./fonts/[name].[ext]",
+                            publicPath: "/",
+                            esModule: false,
+                        },
+                    },
+                },
+                {
+                    test: /\.(jpe?g|png|bmp|gif)$/,
+                    exclude: /node_modules/,
+                    use: {
+                        loader: "file-loader",
+                        options: {
+                            name: "./images/[name].[ext]",
+                            publicPath: "/",
+                            esModule: false,
+                        },
                     },
                 },
             ],
@@ -39,9 +91,13 @@ module.exports = (env, argv) => {
             new WebpackManifestPlugin(),
             new BundleAnalyzerPlugin({
                 analyzerMode: argv.mode === "development" ? "server" : "static",
+                openAnalyzer: argv.mode === "development",
+            }),
+            new MiniCssExtractPlugin({
+                filename: "[name].css",
             }),
         ],
-        devtool: argv.mode === "development" ? "eval" : "hidden-source-map",
+        devtool: argv.mode === "development" ? "eval" : "source-map",
         devServer: {
             contentBase: output,
         },