+++ src/components/CTAElement/index.js
@@ -1,0 +1,3 @@
+import { CTAElement } from "./CTAElement.component";
+
+export { CTAElement };