+++ src/components/Copyright/index.js
@@ -1,0 +1,3 @@
+import { Copyright } from "./Copyright.component";
+
+export { Copyright };