+++ src/components/Navigation/index.js
@@ -1,0 +1,3 @@
+import { Navigation } from "./Navigation.component";
+
+export { Navigation };